checkout functionallity works SAVE OR REVERT TO HERE IF BROKENga .

+++ cart/context_processor.py
@@ -5,7 +5,7 @@ def cart_total_amount(request):
 		cart = Cart(request)
 		total_bill = 0.0
 		for key,value in request.session['cart'].items():
-			total_bill = total_bill + (float(value['price']) * value['quantity'])
+			total_bill = total_bill + (int(value['price']) * value['quantity'])
 		return {'cart_total_amount' : total_bill} 
 	else:
 		return {'cart_total_amount' : 0}
